refactor: route-drive the users tab header actions
Select each tab's header actions by route (a <Routes> block) instead of a
binary conditional, mirroring how the tab bodies are already routed. Extract
the Users tab's actions into a self-contained UsersHeaderActions component.
Adding a future tab is now: one tabs[] entry plus one <Route> in each of the
actions/mobile-search/body <Routes> blocks, with no shared conditionals.
