+++ frontend/src/component/admin/users/UsersTabs/UsersTabs.tsx
@@ -1,29 +1,19 @@
 import { useState } from 'react';
-import {
-    Button,
-    IconButton,
-    Tab,
-    Tabs,
-    Tooltip,
-    styled,
-    useMediaQuery,
-} from '@mui/material';
-import Download from '@mui/icons-material/Download';
-import { Route, Routes, useLocation, useNavigate } from 'react-router';
+import { Tab, Tabs, styled, useMediaQuery } from '@mui/material';
+import { Route, Routes, useLocation } from 'react-router';
 import { ADMIN } from 'component/providers/AccessProvider/permissions';
 import { PermissionGuard } from 'component/common/PermissionGuard/PermissionGuard';
 import { PageContent } from 'component/common/PageContent/PageContent';
-import { PageHeader } from 'component/common/PageHeader/PageHeader';
 import { Search } from 'component/common/Search/Search';
 import { TabLink } from 'component/common/TabNav/TabLink';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { PremiumFeature } from 'component/common/PremiumFeature/PremiumFeature';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
 import { useUsers } from 'hooks/api/getters/useUsers/useUsers';
-import { useAccessOverviewApi } from 'hooks/api/actions/useAccessOverviewApi/useAccessOverviewApi';
 import { usePageTitle } from 'hooks/usePageTitle';
 import theme from 'themes/theme';
 import UsersList from '../UsersList/UsersList.tsx';
+import { UsersHeaderActions } from '../UsersList/UsersHeaderActions.tsx';
 import { InactiveUsersListBody } from '../InactiveUsersList/InactiveUsersListBody.tsx';
 import { InactiveUsersHeaderActions } from '../InactiveUsersList/InactiveUsersHeaderActions.tsx';
 import EditUser from '../EditUser/EditUser.tsx';
@@ -48,16 +38,16 @@ const StyledActions = styled('div')({
 const UsersTabsView = () => {
     usePageTitle('Users');
     const { pathname } = useLocation();
-    const navigate = useNavigate();
     const { isEnterprise } = useUiConfig();
     const { users, loading } = useUsers();
-    const { downloadCSV } = useAccessOverviewApi();
 
     const [searchValue, setSearchValue] = useState('');
 
     const isSmallScreen = useMediaQuery(theme.breakpoints.down('md'));
-    const isInactiveTab = pathname.includes('/inactive');
 
+    // Each tab owns its header actions and body; both are selected by route so
+    // adding another tab is just another entry here and one <Route> in each of
+    // the <Routes> blocks below.
     const tabs = [
         {
             label: `Users (${users.length})`,
@@ -101,56 +91,43 @@ const UsersTabsView = () => {
                             </Tabs>
                         </StyledTabsContainer>
                         <StyledActions>
-                            <ConditionallyRender
-                                condition={isInactiveTab}
-                                show={
-                                    <ConditionallyRender
-                                        condition={isEnterprise()}
-                                        show={<InactiveUsersHeaderActions />}
-                                    />
-                                }
-                                elseShow={
-                                    <>
-                                        <ConditionallyRender
-                                            condition={!isSmallScreen}
-                                            show={
-                                                <Search
-                                                    initialValue={searchValue}
-                                                    onChange={setSearchValue}
-                                                />
-                                            }
+                            <Routes>
+                                <Route
+                                    path='inactive'
+                                    element={
+                                        isEnterprise() ? (
+                                            <InactiveUsersHeaderActions />
+                                        ) : null
+                                    }
+                                />
+                                <Route
+                                    path='*'
+                                    element={
+                                        <UsersHeaderActions
+                                            searchValue={searchValue}
+                                            onSearch={setSearchValue}
+                                            isSmallScreen={isSmallScreen}
                                         />
-                                        <PageHeader.Divider />
-                                        <Tooltip
-                                            title='Exports user access information'
-                                            arrow
-                                            describeChild
-                                        >
-                                            <IconButton onClick={downloadCSV}>
-                                                <Download />
-                                            </IconButton>
-                                        </Tooltip>
-                                        <Button
-                                            variant='contained'
-                                            color='primary'
-                                            onClick={() =>
-                                                navigate('/admin/create-user')
-                                            }
-                                        >
-                                            Add new user
-                                        </Button>
-                                    </>
-                                }
-                            />
+                                    }
+                                />
+                            </Routes>
                         </StyledActions>
                     </StyledHeader>
                     <ConditionallyRender
-                        condition={isSmallScreen && !isInactiveTab}
+                        condition={isSmallScreen}
                         show={
-                            <Search
-                                initialValue={searchValue}
-                                onChange={setSearchValue}
-                            />
+                            <Routes>
+                                <Route path='inactive' element={null} />
+                                <Route
+                                    path='*'
+                                    element={
+                                        <Search
+                                            initialValue={searchValue}
+                                            onChange={setSearchValue}
+                                        />
+                                    }
+                                />
+                            </Routes>
                         }
                     />
                 </>
